#158 refactor : divide views and django orm optimization
